# pkg/handlers/handlers.go
// Package handlers contains HTTP request handlers.
package handlers

import (
	"encoding/json"
	"net/http"

	"github.com/go-chi/chi/v5"
)

// Response helpers

func respondJSON(w http.ResponseWriter, status int, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if data != nil {
		json.NewEncoder(w).Encode(data)
	}
}

func respondError(w http.ResponseWriter, status int, message string) {
	respondJSON(w, status, map[string]string{"error": message})
}

// Health endpoints

// Healthz is the liveness probe endpoint.
func Healthz(w http.ResponseWriter, r *http.Request) {
	respondJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

// Readyz is the readiness probe endpoint.
func Readyz(w http.ResponseWriter, r *http.Request) {
	// TODO: Check Kubernetes API connectivity
	respondJSON(w, http.StatusOK, map[string]string{"status": "ready"})
}

// OpenAPISpec serves the OpenAPI specification.
func OpenAPISpec(w http.ResponseWriter, r *http.Request) {
	// TODO: Generate and serve OpenAPI spec
	respondJSON(w, http.StatusOK, map[string]string{
		"openapi": "3.0.0",
		"info": "ECK UI API",
	})
}

// Organization handlers

// ListOrganizations returns all organizations the user has access to.
func ListOrganizations(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement organization listing
	respondJSON(w, http.StatusOK, map[string]interface{}{
		"items": []interface{}{},
	})
}

// GetOrganization returns a single organization.
func GetOrganization(w http.ResponseWriter, r *http.Request) {
	org := chi.URLParam(r, "org")
	// TODO: Implement organization retrieval
	respondJSON(w, http.StatusOK, map[string]string{
		"name": org,
	})
}

// Elasticsearch handlers

// ListElasticsearch returns all Elasticsearch clusters in an organization.
func ListElasticsearch(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement Elasticsearch listing
	respondJSON(w, http.StatusOK, map[string]interface{}{
		"items": []interface{}{},
	})
}

// CreateElasticsearch creates a new Elasticsearch cluster.
func CreateElasticsearch(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement Elasticsearch creation
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// GetElasticsearch returns a single Elasticsearch cluster.
func GetElasticsearch(w http.ResponseWriter, r *http.Request) {
	name := chi.URLParam(r, "name")
	// TODO: Implement Elasticsearch retrieval
	respondJSON(w, http.StatusOK, map[string]string{
		"name": name,
	})
}

// UpdateElasticsearch updates an Elasticsearch cluster.
func UpdateElasticsearch(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement Elasticsearch update
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// DeleteElasticsearch deletes an Elasticsearch cluster.
func DeleteElasticsearch(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement Elasticsearch deletion
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// Kibana handlers

// ListKibana returns all Kibana instances in an organization.
func ListKibana(w http.ResponseWriter, r *http.Request) {
	respondJSON(w, http.StatusOK, map[string]interface{}{"items": []interface{}{}})
}

// CreateKibana creates a new Kibana instance.
func CreateKibana(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// GetKibana returns a single Kibana instance.
func GetKibana(w http.ResponseWriter, r *http.Request) {
	name := chi.URLParam(r, "name")
	respondJSON(w, http.StatusOK, map[string]string{"name": name})
}

// UpdateKibana updates a Kibana instance.
func UpdateKibana(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// DeleteKibana deletes a Kibana instance.
func DeleteKibana(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// APM Server handlers

// ListAPMServer returns all APM Server instances in an organization.
func ListAPMServer(w http.ResponseWriter, r *http.Request) {
	respondJSON(w, http.StatusOK, map[string]interface{}{"items": []interface{}{}})
}

// CreateAPMServer creates a new APM Server instance.
func CreateAPMServer(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// GetAPMServer returns a single APM Server instance.
func GetAPMServer(w http.ResponseWriter, r *http.Request) {
	name := chi.URLParam(r, "name")
	respondJSON(w, http.StatusOK, map[string]string{"name": name})
}

// UpdateAPMServer updates an APM Server instance.
func UpdateAPMServer(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// DeleteAPMServer deletes an APM Server instance.
func DeleteAPMServer(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// Agent handlers

// ListAgent returns all Agent deployments in an organization.
func ListAgent(w http.ResponseWriter, r *http.Request) {
	respondJSON(w, http.StatusOK, map[string]interface{}{"items": []interface{}{}})
}

// CreateAgent creates a new Agent deployment.
func CreateAgent(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// GetAgent returns a single Agent deployment.
func GetAgent(w http.ResponseWriter, r *http.Request) {
	name := chi.URLParam(r, "name")
	respondJSON(w, http.StatusOK, map[string]string{"name": name})
}

// UpdateAgent updates an Agent deployment.
func UpdateAgent(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// DeleteAgent deletes an Agent deployment.
func DeleteAgent(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// Beat handlers

// ListBeat returns all Beat deployments in an organization.
func ListBeat(w http.ResponseWriter, r *http.Request) {
	respondJSON(w, http.StatusOK, map[string]interface{}{"items": []interface{}{}})
}

// CreateBeat creates a new Beat deployment.
func CreateBeat(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// GetBeat returns a single Beat deployment.
func GetBeat(w http.ResponseWriter, r *http.Request) {
	name := chi.URLParam(r, "name")
	respondJSON(w, http.StatusOK, map[string]string{"name": name})
}

// UpdateBeat updates a Beat deployment.
func UpdateBeat(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// DeleteBeat deletes a Beat deployment.
func DeleteBeat(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// Logstash handlers

// ListLogstash returns all Logstash deployments in an organization.
func ListLogstash(w http.ResponseWriter, r *http.Request) {
	respondJSON(w, http.StatusOK, map[string]interface{}{"items": []interface{}{}})
}

// CreateLogstash creates a new Logstash deployment.
func CreateLogstash(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// GetLogstash returns a single Logstash deployment.
func GetLogstash(w http.ResponseWriter, r *http.Request) {
	name := chi.URLParam(r, "name")
	respondJSON(w, http.StatusOK, map[string]string{"name": name})
}

// UpdateLogstash updates a Logstash deployment.
func UpdateLogstash(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}

// DeleteLogstash deletes a Logstash deployment.
func DeleteLogstash(w http.ResponseWriter, r *http.Request) {
	respondError(w, http.StatusNotImplemented, "not implemented")
}